// src/app/core/Interfaces/upload-movie/IUploadMovieInput.ts
/**
 * Iupload movie input
 */
export interface IUploadMovieInput {
  /**
   * id
   */
  id?: string;
  /**
   * name
   */
  name?: string;
  /**
   * title
   *
   */
  title?: string;
  /**
   * description
   */
  description?: string;
  /**
   * language
   */
  language?: string;
  /**
   * genre
   */
  genre?: string;
  /**
   * imdnRating
   */
  imdnRating?: string;
  /**
   * imageUrl
   */
  imageUrl?: string;
  /**
   * type
   */
  type?: string;
  /**
   * isPrime
   */
  isPrime?: boolean;
}